pyproject.toml
[tool.poetry]
name = "flaskgpt"
version = "0.1.0"
description = "ChatGPT Clone using Flask"
authors = ["KJD <125159447+Thrylos77@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "2.3.3"
openai = "^1.109.0"
python-dotenv = "1.0.0"
bleach = "^6.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
